--- a/core/src/main/java/io/questdb/cairo/CairoConfiguration.java
+++ b/core/src/main/java/io/questdb/cairo/CairoConfiguration.java
@@ -594,6 +594,8 @@ public interface CairoConfiguration {
 
     boolean isSqlJitDebugEnabled();
 
+    boolean isSqlOrderByRadixSortEnabled();
+
     boolean isSqlParallelFilterEnabled();
 
     boolean isSqlParallelFilterPreTouchEnabled();