--- a/core/src/main/java/io/questdb/cairo/CairoConfiguration.java
+++ b/core/src/main/java/io/questdb/cairo/CairoConfiguration.java
@@ -435,6 +435,8 @@ public interface CairoConfiguration {
 
     int getSqlModelPoolCapacity();
 
+    int getSqlOrderByRadixSortThreshold();
+
     int getSqlPageFrameMaxRows();
 
     int getSqlPageFrameMinRows();
@@ -594,7 +596,7 @@ public interface CairoConfiguration {
 
     boolean isSqlJitDebugEnabled();
 
-    boolean isSqlOrderByRadixSortEnabled();
+    boolean isSqlOrderBySortEnabled();
 
     boolean isSqlParallelFilterEnabled();