--- a/core/src/main/java/io/questdb/cairo/DefaultCairoConfiguration.java
+++ b/core/src/main/java/io/questdb/cairo/DefaultCairoConfiguration.java
@@ -764,6 +764,11 @@ public class DefaultCairoConfiguration implements CairoConfiguration {
     }
 
     @Override
+    public int getSqlOrderByRadixSortThreshold() {
+        return 600;
+    }
+
+    @Override
     public int getSqlPageFrameMaxRows() {
         return 1_000_000;
     }
@@ -1103,7 +1108,7 @@ public class DefaultCairoConfiguration implements CairoConfiguration {
     }
 
     @Override
-    public boolean isSqlOrderByRadixSortEnabled() {
+    public boolean isSqlOrderBySortEnabled() {
         return true;
     }