--- a/core/src/main/java/io/questdb/cairo/sql/DelegatingRecordCursor.java
+++ b/core/src/main/java/io/questdb/cairo/sql/DelegatingRecordCursor.java
@@ -24,8 +24,9 @@
 
 package io.questdb.cairo.sql;
 
+import io.questdb.griffin.SqlException;
 import io.questdb.griffin.SqlExecutionContext;
 
 public interface DelegatingRecordCursor extends RecordCursor {
-    void of(RecordCursor base, SqlExecutionContext executionContext);
+    void of(RecordCursor base, SqlExecutionContext executionContext) throws SqlException;
 }