--- a/core/src/main/java/io/questdb/griffin/engine/LimitRecordCursorFactory.java
+++ b/core/src/main/java/io/questdb/griffin/engine/LimitRecordCursorFactory.java
@@ -30,6 +30,7 @@ import io.questdb.cairo.sql.*;
 import io.questdb.griffin.PlanSink;
 import io.questdb.griffin.SqlException;
 import io.questdb.griffin.SqlExecutionContext;
+import io.questdb.std.Misc;
 import org.jetbrains.annotations.Nullable;
 
 public class LimitRecordCursorFactory extends AbstractRecordCursorFactory {
@@ -56,13 +57,13 @@ public class LimitRecordCursorFactory extends AbstractRecordCursorFactory {
         final RecordCursor baseCursor = base.getCursor(executionContext);
         try {
             cursor.of(baseCursor, executionContext);
+            return cursor;
         } catch (Throwable th) {
-            baseCursor.close();
+            cursor.close();
             throw th;
         } finally {
             executionContext.setColumnPreTouchEnabled(preTouchEnabled);
         }
-        return cursor;
     }
 
     @Override
@@ -148,7 +149,7 @@ public class LimitRecordCursorFactory extends AbstractRecordCursorFactory {
 
         @Override
         public void close() {
-            base.close();
+            base = Misc.free(base);
         }
 
         @Override