--- a/core/src/main/java/io/questdb/griffin/engine/groupby/CountRecordCursorFactory.java
+++ b/core/src/main/java/io/questdb/griffin/engine/groupby/CountRecordCursorFactory.java
@@ -57,7 +57,7 @@ public class CountRecordCursorFactory extends AbstractRecordCursorFactory {
             cursor.of(baseCursor, executionContext.getCircuitBreaker());
             return cursor;
         } catch (Throwable th) {
-            baseCursor.close();
+            cursor.close();
             throw th;
         }
     }