--- a/core/src/main/java/io/questdb/griffin/engine/groupby/VirtualFunctionSkewedSymbolRecordCursor.java
+++ b/core/src/main/java/io/questdb/griffin/engine/groupby/VirtualFunctionSkewedSymbolRecordCursor.java
@@ -31,7 +31,6 @@ import io.questdb.std.Misc;
 import io.questdb.std.ObjList;
 
 public class VirtualFunctionSkewedSymbolRecordCursor extends AbstractVirtualFunctionRecordCursor {
-
     protected RecordCursor managedCursor;
 
     public VirtualFunctionSkewedSymbolRecordCursor(ObjList<Function> functions) {