Introduces a quick or radix sort-based factory used for unlimited ORDER BY over a single INT / IPv4 / LONG / TIMESTAMP / DATE column or expression. The memory overhead is similar to the existing `SortedRecordCursorFactory`: each row takes 32 bytes vs 36 bytes in the red-black tree-based factory (the latter assumes #4888).

### Benchmarks

The following numbers were obtained on i7-1185G7 CPU with 32GB RAM, running Ubuntu 22.04 64-bit and OpenJDK 17.0.12. The dataset is the one from ClickBench.

```sql
-- 99M rows
-- master: 320s
-- radix sort: 4s
SELECT * FROM hits ORDER BY UserID;

-- 10M rows
-- master: 14s
-- radix sort: 400ms
SELECT * FROM (hits LIMIT 10000000) ORDER BY UserID;

-- 1M rows
-- master: 456ms
-- radix sort: 50ms
SELECT * FROM (hits LIMIT 1000000) ORDER BY UserID;

-- 10K rows
-- master: 12ms
-- radix sort: 9ms
SELECT * FROM (hits LIMIT 10000) ORDER BY UserID;

-- 100 rows
-- master: 5ms
-- radix sort: 3ms
SELECT * FROM (hits LIMIT 100) ORDER BY UserID;
```