@ideoma 2024-08-23
#### [PR Coverage check]

:heart_eyes: **pass** : 125 / 134 (93.28%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|io/questdb/griffin/engine/orderby/LongSortedLightRecordCursorFactory.java|26|31|83.87%|
|:large_blue_circle:|io/questdb/griffin/engine/orderby/LongSortedLightRecordCursor.java|82|86|95.35%|
|:large_blue_circle:|io/questdb/PropServerConfiguration.java|4|4|100.00%|
|:large_blue_circle:|io/questdb/cairo/DefaultCairoConfiguration.java|2|2|100.00%|
|:large_blue_circle:|io/questdb/PropertyKey.java|2|2|100.00%|
|:large_blue_circle:|io/questdb/griffin/SqlCodeGenerator.java|7|7|100.00%|
|:large_blue_circle:|io/questdb/cairo/CairoConfigurationWrapper.java|2|2|100.00%|