@ideoma 2024-08-24
#### [PR Coverage check]

:heart_eyes: **pass** : 131 / 140 (93.57%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|io/questdb/griffin/engine/orderby/LongSortedLightRecordCursorFactory.java|27|32|84.38%|
|:large_blue_circle:|io/questdb/griffin/engine/orderby/LongSortedLightRecordCursor.java|86|90|95.56%|
|:large_blue_circle:|io/questdb/PropServerConfiguration.java|4|4|100.00%|
|:large_blue_circle:|io/questdb/cairo/DefaultCairoConfiguration.java|2|2|100.00%|
|:large_blue_circle:|io/questdb/PropertyKey.java|2|2|100.00%|
|:large_blue_circle:|io/questdb/griffin/SqlCodeGenerator.java|8|8|100.00%|
|:large_blue_circle:|io/questdb/cairo/CairoConfigurationWrapper.java|2|2|100.00%|