@ideoma 2024-08-27
#### [PR Coverage check]

:heart_eyes: **pass** : 222 / 266 (83.46%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|io/questdb/griffin/engine/table/SelectedRecordCursorFactory.java|0|1|00.00%|
|:large_blue_circle:|io/questdb/griffin/engine/groupby/FillRangeRecordCursorFactory.java|0|1|00.00%|
|:large_blue_circle:|io/questdb/griffin/engine/orderby/SortedRecordCursorFactory.java|0|1|00.00%|
|:large_blue_circle:|io/questdb/griffin/engine/groupby/CountRecordCursorFactory.java|0|1|00.00%|
|:large_blue_circle:|io/questdb/griffin/engine/groupby/SampleByInterpolateRecordCursorFactory.java|2|7|28.57%|
|:large_blue_circle:|io/questdb/griffin/engine/orderby/SortedLightRecordCursorFactory.java|1|3|33.33%|
|:large_blue_circle:|io/questdb/griffin/engine/orderby/LimitedSizeSortedLightRecordCursorFactory.java|5|11|45.45%|
|:large_blue_circle:|io/questdb/griffin/engine/groupby/GroupByRecordCursorFactory.java|5|8|62.50%|
|:large_blue_circle:|io/questdb/griffin/engine/groupby/DistinctRecordCursorFactory.java|2|3|66.67%|
|:large_blue_circle:|io/questdb/griffin/engine/LimitRecordCursorFactory.java|2|3|66.67%|
|:large_blue_circle:|io/questdb/griffin/engine/groupby/AbstractSampleByFillRecordCursorFactory.java|6|9|66.67%|
|:large_blue_circle:|io/questdb/griffin/engine/groupby/AbstractSampleByRecordCursorFactory.java|6|9|66.67%|
|:large_blue_circle:|io/questdb/griffin/engine/join/RecordAsAFieldRecordCursorFactory.java|8|11|72.73%|
|:large_blue_circle:|io/questdb/griffin/engine/groupby/DistinctTimeSeriesRecordCursorFactory.java|4|5|80.00%|
|:large_blue_circle:|io/questdb/griffin/engine/orderby/RecordTreeChain.java|4|5|80.00%|
|:large_blue_circle:|io/questdb/griffin/engine/orderby/LimitedSizePartiallySortedLightRecordCursor.java|9|11|81.82%|
|:large_blue_circle:|io/questdb/griffin/engine/table/LatestByRecordCursorFactory.java|5|6|83.33%|
|:large_blue_circle:|io/questdb/griffin/engine/orderby/LongSortedLightRecordCursorFactory.java|23|27|85.19%|
|:large_blue_circle:|io/questdb/griffin/engine/orderby/LongSortedLightRecordCursor.java|86|90|95.56%|
|:large_blue_circle:|io/questdb/griffin/engine/groupby/SampleByFillNoneRecordCursorFactory.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/orderby/SortedRecordCursor.java|9|9|100.00%|
|:large_blue_circle:|io/questdb/PropServerConfiguration.java|4|4|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/groupby/AbstractSampleByNotKeyedRecordCursorFactory.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/orderby/LimitedSizeSortedLightRecordCursor.java|10|10|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/orderby/SortedLightRecordCursor.java|11|11|100.00%|
|:large_blue_circle:|io/questdb/cairo/DefaultCairoConfiguration.java|2|2|100.00%|
|:large_blue_circle:|io/questdb/PropertyKey.java|2|2|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/groupby/AbstractNoRecordSampleByCursor.java|3|3|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/table/SelectedRecordCursor.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/SqlCodeGenerator.java|8|8|100.00%|
|:large_blue_circle:|io/questdb/cairo/CairoConfigurationWrapper.java|2|2|100.00%|